hibernate: avoid the data in hidden area to be snapshotted
The sensitive data in hidden area should not be snapshotted by
hibernate. On the other hand, the hidden area in boot kernel
should be cleaned before the system be restored to target kernel.

This patch only be used by hibernation verification. This patch will
also be replaced after hibernation encryption is ready on mainline.

Link: https://lkml.org/lkml/2019/1/3/281