PM / hibernate: Generate and verify signature for snapshot image
When producing memory snapshot image, hibernation uses HMAC-SHA512
with EFI secret key to calculate the hash for all data pages in image.
The hash result will be kept in the snapshot header as the image
signature. Before hibernation restores system, kernel executes
HMAC-SHA512 again and compares the result with the signature in header
to verify the integrity of snapshot image.

If the verification failed, the resume process will be stopped. The
snapshot image will be discarded and system will boot as normal.

This patch will be replaced after hibernation encryption with TPM
is ready on mainline. It needs some big changes after review by
upstream:
- Simplify the design: remove keyring dependency and trampoline.
- Encrypted whole snapshot image instead of only data pages.
- Using TPM:
	- Direct use TPM API in hibernation instead of keyring
        - Localities (suggested by James Bottomley)

Link: https://lkml.org/lkml/2019/1/3/281