PM / hibernate: a option to request that snapshot image must be validly signed
This kernel option is similar with the option for kernel module signature
verification. When the option is unselected, kernel will be tainted by
restoring from a unsigned snapshot image or the image was signed by a
wrong key.

When the option is selected, kernel will refuse the system to be resumed
if the signature verification did not pass. The hibernation resume
process will be stopped and whole system just boots as normal.

This patch will be replaced after hibernation encryption with TPM
is ready on mainline. It needs some big changes after review by
upstream:
- Simplify the design: remove keyring dependency and trampoline.
- Encrypted whole snapshot image instead of only data pages.
- Using TPM:
	- Direct use TPM API in hibernation instead of keyring
        - Localities (suggested by James Bottomley)

Link: https://lkml.org/lkml/2019/1/3/281