PM / hibernate: require hibernate snapshot image to be validly signed when kernel is locked down
When the kernel is locked down, the snapshot must be signed by secret
key. Kernel should stops whole resuming process when the snapshot is
unsigned or signed by wrong key. If the efi secret key is invalid
(when secure boot is disabled), the hibernation function will be locked
down.

This patch will be replaced after hibernation encryption with TPM
is ready on mainline. It needs some big changes after review by
upstream:
- Simplify the design: remove keyring dependency and trampoline.
- Encrypted whole snapshot image instead of only data pages.
- Using TPM:
        - Direct use TPM API in hibernation instead of keyring
        - Localities (suggested by James Bottomley)

Link: https://lkml.org/lkml/2019/1/3/281