PM / hibernate: let hibernate verification depends on force lock down
Hibernate verification should depends on kerne lock down force
integrity mode or force confidentiality mode. Because that the
EFI secret key in kernel space can not be modified at runtime.

On the other hand, because EFI secret key be loaded in early boot
stage. So the kernel lock down must be built as _force_ mode to
protect it in early boot stage.